+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+### Fixed
+
+- Incorrect parent class when extending from custom interactors
+
 ## [1.3.1] - 2019-05-23
 
 ### Fixed