+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ### Fixed
 
+- Non-enumerable static properties not being carried over
 - Incorrect parent class when extending from custom interactors
 
 ## [1.3.1] - 2019-05-23